infra/backup-job/requirements.txt
azure-storage-blob==12.23.1
azure-identity==1.17.1
